build(WriteVersionToBuildLog): Fixed build
The WriteVersionToBuildLog target is now only called on the
Singer.csproj project, not on any others.
This should be fine as we just want to update ADO's version numbers.
Setting the actual assembly version numbers can be controlled by whether
the GitVersionTask nuget packages is added to a specific project or not.