fix: Fixed duplicate user creation
Fixed 2 problems:
Because of wrong AutoMapper configuration (for all users) and incomplete
Queryable members in LegalGuardianUserService, user properties would get
lost during an update, resulting in duplicate user creation when
updating.
In the CreateAsync method of UserService, a call was made to
UserManager.AddAsync, but then another call to base.AddAsync with the
DTO, so the same user was created twice. This is now done manually in
UserService with the context of the entity created by UserManager