fix: Fixed event duration logic
The event duration logic in the frontend had 2 issues:
The startDateTime and endDateTime were not Date types but strings. There
was some ugly code doing HTTP calls inside the component, so I moved
this into a more suited place into the event service. There we now
perform correct type conversion to Date for the relevant fields.
Then there was an issue in checking whether the event lasted a day or
more, probably a leftover from when we had startDate and endDate
properties. This now currectly gets the year/month/day from the date
type, and compares them, ignoring the hours/minutes in the process.

Closes AB#107