fix(UserService): Corrected check for existing users
The previous check checked on the specific user type (e.g. AdminUsers),
however if a user with the same email address existed in another table
(e.g. LegalGuardianUsers) the check would not find that user and proceed
with trying to add the new admin user. This would fail with a vague
error